Rewrite client and move it to network/ dir
